+++ app/(home)/_layout.tsx
@@ -1,0 +1,5 @@
+import { Slot, Stack } from "expo-router";
+
+export default function HomeLayout(){
+ return <Slot />   
+}
